requirements.txt
-i https://pypi.org/simple
aiohappyeyeballs==2.6.1; python_version >= '3.9'
aiohttp==3.12.15; python_version >= '3.9'
aiosignal==1.4.0; python_version >= '3.9'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.10.0; python_version >= '3.9'
attrs==25.3.0; python_version >= '3.8'
backoff==2.2.1; python_version >= '3.7' and python_version < '4.0'
bcrypt==4.3.0; python_version >= '3.8'
build==1.3.0; python_version >= '3.9'
cachetools==5.5.2; python_version >= '3.7'
certifi==2025.8.3; python_version >= '3.7'
charset-normalizer==3.4.2; python_version >= '3.7'
chromadb==1.0.15; python_version >= '3.9'
click==8.2.1; python_version >= '3.10'
coloredlogs==15.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
dataclasses-json==0.6.7; python_version >= '3.7' and python_version < '4.0'
distro==1.9.0; python_version >= '3.6'
durationpy==0.10
filelock==3.18.0; python_version >= '3.9'
flatbuffers==25.2.10
frozenlist==1.7.0; python_version >= '3.9'
fsspec==2025.7.0; python_version >= '3.9'
google-auth==2.40.3; python_version >= '3.7'
googleapis-common-protos==1.70.0; python_version >= '3.7'
grpcio==1.74.0; python_version >= '3.9'
h11==0.16.0; python_version >= '3.8'
hf-xet==1.1.5; python_version >= '3.8'
httpcore==1.0.9; python_version >= '3.8'
httptools==0.6.4; python_full_version >= '3.8.0'
httpx==0.28.1; python_version >= '3.8'
httpx-sse==0.4.1; python_version >= '3.9'
huggingface-hub==0.34.3; python_full_version >= '3.8.0'
humanfriendly==10.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
idna==3.10; python_version >= '3.6'
importlib-metadata==8.7.0; python_version >= '3.9'
importlib-resources==6.5.2; python_version >= '3.9'
jiter==0.10.0; python_version >= '3.9'
jsonpatch==1.33; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
jsonpointer==3.0.0; python_version >= '3.7'
jsonschema==4.25.0; python_version >= '3.9'
jsonschema-specifications==2025.4.1; python_version >= '3.9'
kubernetes==33.1.0; python_version >= '3.6'
langchain==0.3.27; python_version >= '3.9' and python_version < '4.0'
langchain-chroma==0.2.5; python_version >= '3.9'
langchain-community==0.3.27; python_version >= '3.9'
langchain-core==0.3.72; python_version >= '3.9'
langchain-openai==0.3.28; python_version >= '3.9'
langchain-text-splitters==0.3.9; python_version >= '3.9'
langsmith==0.4.11; python_version >= '3.9'
markdown-it-py==3.0.0; python_version >= '3.8'
marshmallow==3.26.1; python_version >= '3.9'
mdurl==0.1.2; python_version >= '3.7'
mmh3==5.2.0; python_version >= '3.9'
mpmath==1.3.0
multidict==6.6.3; python_version >= '3.9'
mypy-extensions==1.1.0; python_version >= '3.8'
numpy==2.3.2; python_version >= '3.11'
oauthlib==3.3.1; python_version >= '3.8'
onnxruntime==1.22.1; python_version >= '3.10'
openai==1.99.1; python_version >= '3.8'
opentelemetry-api==1.36.0; python_version >= '3.9'
opentelemetry-exporter-otlp-proto-common==1.36.0; python_version >= '3.9'
opentelemetry-exporter-otlp-proto-grpc==1.36.0; python_version >= '3.9'
opentelemetry-proto==1.36.0; python_version >= '3.9'
opentelemetry-sdk==1.36.0; python_version >= '3.9'
opentelemetry-semantic-conventions==0.57b0; python_version >= '3.9'
orjson==3.11.1; python_version >= '3.9'
overrides==7.7.0; python_version >= '3.6'
packaging==25.0; python_version >= '3.8'
posthog==5.4.0; python_version >= '3.9'
propcache==0.3.2; python_version >= '3.9'
protobuf==6.31.1; python_version >= '3.9'
pyasn1==0.6.1; python_version >= '3.8'
pyasn1-modules==0.4.2; python_version >= '3.8'
pybase64==1.4.2; python_version >= '3.8'
pydantic==2.11.7; python_version >= '3.9'
pydantic-core==2.33.2; python_version >= '3.9'
pydantic-settings==2.10.1; python_version >= '3.9'
pygments==2.19.2; python_version >= '3.8'
pypika==0.48.9
pyproject-hooks==1.2.0; python_version >= '3.7'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.1.1; python_version >= '3.9'
pyyaml==6.0.2; python_version >= '3.8'
referencing==0.36.2; python_version >= '3.9'
regex==2025.7.34; python_version >= '3.9'
requests==2.32.4; python_version >= '3.8'
requests-oauthlib==2.0.0; python_version >= '3.4'
requests-toolbelt==1.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rich==14.1.0; python_full_version >= '3.8.0'
rpds-py==0.26.0; python_version >= '3.9'
rsa==4.9.1; python_version >= '3.6' and python_version < '4'
shellingham==1.5.4; python_version >= '3.7'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.42; python_version >= '3.7'
sympy==1.14.0; python_version >= '3.9'
tenacity==9.1.2; python_version >= '3.9'
tiktoken==0.10.0; python_version >= '3.9'
tokenizers==0.21.4; python_version >= '3.9'
tqdm==4.67.1; python_version >= '3.7'
typer==0.16.0; python_version >= '3.7'
typing-extensions==4.14.1; python_version >= '3.9'
typing-inspect==0.9.0
typing-inspection==0.4.1; python_version >= '3.9'
urllib3==2.5.0; python_version >= '3.9'
uvicorn[standard]==0.35.0; python_version >= '3.9'
uvloop==0.21.0; python_full_version >= '3.8.0'
watchfiles==1.1.0; python_version >= '3.9'
websocket-client==1.8.0; python_version >= '3.8'
websockets==15.0.1; python_version >= '3.9'
yarl==1.20.1; python_version >= '3.9'
zipp==3.23.0; python_version >= '3.9'
zstandard==0.23.0; python_version >= '3.8'
